// client/src/view/classNames/darkThemeClassNames.ts
enum darkThemeClassNames {
  calculatorDark = 'calculator-dark',
  engineerOptionsDark = 'engineer-options-dark',
  historyDark = 'history-dark',
  inputFieldDark = 'input-field-dark',
  resultOutputDark = 'result-output-dark',
  controlButtonsDark = 'control-buttons-dark',
  mainButtonDark = 'main-button-dark',
  operationsBackgroundDark = 'operations-background-dark',
  operationButtonsDark = 'operation-dark',
  equalDark = 'equals-button-dark',
  numbersDark = 'numbers-dark',
  historyContainerDark = 'history-container-dark',
  historyItemDark = 'history-item-dark',
}

export default darkThemeClassNames;